Remove unused mlflow.db file and update README with task list and backend setup instructions. Refactor Dockerfiles for backend and pipeline, and adjust import paths in FastAPI app. Modify prediction input schema to use floats instead of ints, and enhance prediction service to load label maps from MLflow artifacts. Update orchestration flow to implement a training pipeline with logging for model metrics and label mapping.

+++ backend/app/schemas/predict_input.py
@@ -1,8 +1,8 @@
 from pydantic import BaseModel
 
 class PredictInput(BaseModel):
-    sets_per_match: int
-    receives_per_match: int
-    blocks_per_match: int
-    digs_per_match: int
-    attacks_per_match: int
+    sets_per_match: float
+    receives_per_match: float
+    blocks_per_match: float
+    digs_per_match: float
+    attacks_per_match: float
